fix bot with parallel code
I did not succeed to have a background task (delay) that can wait for other
(parallel) subtask to complete. It seems that calling get() inside a
background task is forbidden.
For now, the slack bot will be synchroneous, this mean the user will not see
a message until the processing is complete. This also means that we keep the
connexion between slack and us open during all this time (~10s) but
apparently slack do not cut it.