Lighten Slack output to show the most important info
If a PR is :x: we don't need to know how many people
raised hand or said LGTM.
If there is some LGTM the :hand: are not relevant.
If there is a lot of :hand: and nothing else, it's
a sign same something is going wrong.

As said in our GitHub Workflow we shouldn't use :hand:
this doesn't really means anything and doesn't have
a real weight on the decision to merge. But as some
people are using it right now, we need to keep that
for a while.