use review API
Review are grouped by user, the idea is to have something similar to what
github provide at the end of the PR page, ie only the last review from a
user is taken into account.
In the same spirit, I also removed the "weak" review notion. It could be
mapped on the COMMENTED status, but it don't bring much value and the
same information can be found with "nb review" (the latter also count
single comment)