cleanup prfetcher
We now use clear classes instead of dict
prfetcher.py is now only responsible to manage celery task.
The url and data are now in githubdata.py

Note that we now use pickle for serialisation instead of json.
Internet warn about security issue but this is the default
The attempt of writting a json serializer result in strange bugs